requirement.txt
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
anyio==4.1.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
Deprecated==1.2.14
Django==4.2.6
django-filter==23.3
django-whatsapp-business-api-is==0.1.25
djangorestframework==3.14.0
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.4
Markdown==3.5
multidict==6.0.4
nexmo==2.5.2
pycparser==2.21
pydantic==1.10.13
PyJWT==2.8.0
python-telegram-bot==20.6
pytz==2023.3.post1
requests==2.31.0
sniffio==1.3.0
sqlparse==0.4.4
telepot==12.7
twilio==8.10.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
vonage==3.11.0
vonage-jwt==1.1.0
wrapt==1.15.0
yarl==1.9.2